Fix rendering issue
When running on my machine the drop-down menu would not render properly
without having a fixed width.
To allow it to be flexible but remain constrained, I used a
ConstrainedBox widget with a maximum box constraint of 240 pixels. This
can be changed later if we need to make the box bigger or if we change
the design