Begin restructuring of crypto functions
Asymmetric key cryptography functions are being restructured into
classes that are held by an InheritedWidget high in the tree. This way
the widgets deep in the tree can call the functions through the
inherited context anywhere.

Need to finish incorporation into working pages and add some code to
test functionality.

TODO: manipulate storage in a similar way so that it can be accessed in
the widget tree and used to store/retreive the asymmetric key values as
necessary.